Add react components and design details
