+++ package.json
@@ -46,7 +46,8 @@
     "prettier": "^2.6.2",
     "ts-node": "^10.9.2",
     "typescript": "^5.3.3",
-    "vite": "^5.1.0"
+    "vite": "^5.1.0",
+    "vite-plugin-svgr": "^4.2.0"
   },
   "packageManager": "yarn@3.2.1"
 }
