+++ src/App.css
@@ -40,3 +40,41 @@
 .read-the-docs {
   color: #888;
 }
+
+.site-logo {
+  width: 100%;
+  margin-left: 2rem;
+}
+
+.site-header {
+  z-index: 99;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 3%;
+}
+
+.site-header--small {
+  padding: var(--size);
+}
+
+.menu {
+  display: flex;
+  list-style: none;
+  margin: 0;
+  padding: 0;
+}
+
+.menu--header {
+  flex-grow: 1;
+  justify-content: space-between;
+}
+
+.menu-item {
+  display: flex;
+  align-items: center;
+}
+
+.menu-item h2 {
+  margin-right: 2rem;
+}
