+++ src/components/Accordion.tsx
@@ -1,0 +1,35 @@
+import { useState } from 'react'
+import './Accordion.css'
+import { replaceText } from '../App'
+
+interface AccordionInterface {
+  title: string
+  content: string | JSX.Element | object
+}
+
+export const Accordion = ({ title, content }: AccordionInterface) => {
+  const [isActive, setIsActive] = useState(false)
+
+  return (
+    <div className="accordion-item">
+      <div className="accordion-title" onClick={() => setIsActive(!isActive)}>
+        <div>{title}</div>
+        <div>{isActive ? '-' : '+'}</div>
+      </div>
+      {isActive && <div className="accordion-draw">{
+        Object.entries(content).map(r => {
+          const text = replaceText(r[0])
+          if (Object.values(r[1])[0] === undefined) {
+            return (
+              <div className="accordion-content">
+                <a href="?text">{text}</a>
+              </div>
+            )
+          } else {
+            return <Accordion title={text} content={r[1]} />
+          }
+        })
+      }</div>}
+    </div>
+  )
+}
