+++ src/index.css
@@ -2,6 +2,7 @@
   margin: 0;
   padding: 0;
   box-sizing: border-box;
+  color: #f9faf8;
 }
 
 body {
@@ -16,38 +17,8 @@ h1 {
   margin: 2rem 0 4rem 0;
 }
 
-.accordion {
-  max-width: 600px;
-  margin: 2rem auto;
-}
-
-.accordion-title {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  cursor: pointer;
-  background-color: #21aeca;
-}
-
-.accordion-title:hover {
-  background-color: #3ab4cc;
-}
-
-.accordion-title,
-.accordion-content {
-  padding: 1rem;
-}
-
-.accordion-content {
-  background-color: #39b9d2;
-}
-
 @media screen and (max-width: 700px) {
   body {
     font-size: 18px;
-  }
-
-  .accordion {
-    width: 90%;
   }
 }
